WB_Scripts/requirements.txt
django-ckeditor==6.5.1
django-filter==23.2
Django==4.2.2
djangorestframework==3.14.0
psycopg2==2.9.6
uwsgi==2.0.21